give sound update back
