+++ supervisor/bootstrap.py
@@ -40,7 +40,7 @@ from .discovery import Discovery
 from .hardware.module import HardwareManager
 from .hassos import HassOS
 from .homeassistant.module import HomeAssistant
-from .host import HostManager
+from .host.manager import HostManager
 from .ingress import Ingress
 from .misc.filter import filter_data
 from .misc.scheduler import Scheduler
